feat(cli): SMI-5129 — wrap author/ subcommand handlers with telemetry
Wraps the 6 author handlers with withTelemetry (namespaced skillIds under the
`author` parent command):
- init/validate/publish: extracted to a new author/init.action.ts sibling
  (init.ts was 490 LOC; in-file wrapping would breach the 500 gate). init.ts
  keeps the logic (initSkill/validateSkill/publishSkill); the factories +
  wrapped actions move to init.action.ts (one-way import, no cycle). index.ts
  re-exports the factories from init.action.js; init.test.ts updated to import
  createInitCommand from the new location.
- mcp-init, subagent, transform: in-file wraps.

CLI_DISPATCHER_MAP coverage 36 → 42. Sizes: init.ts 398, init.action.ts 151,
mcp-init 244, subagent 216, transform 181 (all <500). init.test.ts (15) +
coverage test green.
