+++ packages/cli/src/commands/author/transform.ts
@@ -6,6 +6,7 @@
 
 import { Command } from 'commander'
 import chalk from 'chalk'
+import { withTelemetry } from '@skillsmith/core/telemetry'
 import ora from 'ora'
 import { readFile, readdir } from 'fs/promises'
 import { join, resolve } from 'path'
@@ -139,6 +140,31 @@ export async function transformSkill(skillPath: string, options: TransformOption
   }
 }
 
+// SMI-5129: extracted from inline .action() closure so withTelemetry can wrap it.
+async function transformActionImpl(
+  skillPath: string,
+  opts: Record<string, string | boolean | undefined>
+): Promise<void> {
+  try {
+    await transformSkill(skillPath, {
+      dryRun: opts['dryRun'] as boolean | undefined,
+      force: opts['force'] as boolean | undefined,
+      batch: opts['batch'] as boolean | undefined,
+      tools: opts['tools'] as string | undefined,
+      model: opts['model'] as string | undefined,
+    })
+  } catch (error) {
+    console.error(chalk.red('Error transforming skill:'), sanitizeError(error))
+    process.exit(1)
+  }
+}
+
+export const transformAction = withTelemetry(transformActionImpl, {
+  source: 'cli',
+  extractSkillId: () => 'author transform',
+  extractFramework: () => 'cli',
+})
+
 /**
  * Create transform command
  */
@@ -151,18 +177,5 @@ export function createTransformCommand(): Command {
     .option('--batch', 'Process directory of skills')
     .option('--tools <tools>', 'Override detected tools (comma-separated)')
     .option('--model <model>', 'Model for subagent: sonnet|opus|haiku', 'sonnet')
-    .action(async (skillPath: string, opts: Record<string, string | boolean | undefined>) => {
-      try {
-        await transformSkill(skillPath, {
-          dryRun: opts['dryRun'] as boolean | undefined,
-          force: opts['force'] as boolean | undefined,
-          batch: opts['batch'] as boolean | undefined,
-          tools: opts['tools'] as string | undefined,
-          model: opts['model'] as string | undefined,
-        })
-      } catch (error) {
-        console.error(chalk.red('Error transforming skill:'), sanitizeError(error))
-        process.exit(1)
-      }
-    })
+    .action(transformAction)
 }
